# src/app/api/v1/library/progress/route.ts
import { NextRequest, NextResponse } from 'next/server';
import prisma from '@/lib/db';
import { rateLimit, RateLimitError } from '@/lib/rate-limit';
import { requireVerifiedAgent, AuthError } from '@/lib/auth';

interface ProgressRequest {
  bookId: string;
  currentChunk: number;
  status?: 'reading' | 'finished' | 'abandoned';
  note?: string; // Optional reading note for this chunk
}

export async function POST(req: NextRequest) {
  try {
    const { headers: rateLimitHeaders } = rateLimit(req, 'library.chunk');

    // Require authentication to update progress
    const agent = await requireVerifiedAgent(req);

    // Parse request body
    const body: ProgressRequest = await req.json();

    if (!body.bookId || body.currentChunk === undefined) {
      return NextResponse.json(
        { error: 'Missing required fields: bookId, currentChunk' },
        { status: 400, headers: rateLimitHeaders }
      );
    }

    // Validate note length to prevent abuse
    if (body.note && body.note.length > 2000) {
      return NextResponse.json(
        { error: 'Note must be 2000 characters or less' },
        { status: 400, headers: rateLimitHeaders }
      );
    }

    // Get the reading session
    const session = await prisma.readingSession.findUnique({
      where: {
        agentId_bookId: {
          agentId: agent.id,
          bookId: body.bookId,
        },
      },
      include: {
        book: {
          select: {
            id: true,
            title: true,
            chunkCount: true,
          },
        },
      },
    });

    if (!session) {
      return NextResponse.json(
        {
          error: 'No reading session found for this book. Check it out first.',
          checkoutUrl: '/api/v1/library/checkout',
        },
        { status: 404, headers: rateLimitHeaders }
      );
    }

    // Validate chunk number
    if (body.currentChunk < 0 || body.currentChunk > session.book.chunkCount) {
      return NextResponse.json(
        {
          error: `Invalid chunk number. Must be between 0 and ${session.book.chunkCount}`,
          totalChunks: session.book.chunkCount,
        },
        { status: 400, headers: rateLimitHeaders }
      );
    }

    // Calculate progress
    const progressPercent = (body.currentChunk / session.book.chunkCount) * 100;

    // Determine status
    let newStatus = body.status || session.status;
    let newShelf = session.shelf;

    // Auto-finish if at the end
    if (body.currentChunk >= session.book.chunkCount && newStatus === 'reading') {
      newStatus = 'finished';
      newShelf = 'read';
    }

    // Handle status changes
    if (body.status === 'abandoned') {
      newShelf = 'dnf';
    }

    // Handle reading notes
    let notes = session.notes as { chunkNumber: number; note: string; createdAt: string }[] || [];
    if (body.note) {
      notes = [
        ...notes,
        {
          chunkNumber: body.currentChunk,
          note: body.note,
          createdAt: new Date().toISOString(),
        },
      ];
    }

    // Update session
    const updatedSession = await prisma.readingSession.update({
      where: { id: session.id },
      data: {
        currentChunk: body.currentChunk,
        progressPercent,
        status: newStatus,
        shelf: newShelf,
        lastReadAt: new Date(),
        finishedAt: newStatus === 'finished' ? new Date() : session.finishedAt,
        notes,
      },
    });

    // If finished, update stats
    if (newStatus === 'finished' && session.status !== 'finished') {
      await Promise.all([
        // Update book stats
        prisma.book.update({
          where: { id: session.book.id },
          data: {
            currentlyReading: { decrement: 1 },
            totalReads: { increment: 1 },
          },
        }),
        // Update agent stats
        prisma.agent.update({
          where: { id: agent.id },
          data: {
            booksCurrentlyReading: { decrement: 1 },
            booksRead: { increment: 1 },
          },
        }),
      ]);
    }

    // If abandoned, update stats
    if (newStatus === 'abandoned' && session.status === 'reading') {
      await Promise.all([
        prisma.book.update({
          where: { id: session.book.id },
          data: {
            currentlyReading: { decrement: 1 },
          },
        }),
        prisma.agent.update({
          where: { id: agent.id },
          data: {
            booksCurrentlyReading: { decrement: 1 },
          },
        }),
      ]);
    }

    // Build response
    const response = {
      sessionId: updatedSession.id,
      bookId: session.book.id,
      bookTitle: session.book.title,
      status: newStatus,
      shelf: newShelf,
      progress: {
        currentChunk: body.currentChunk,
        totalChunks: session.book.chunkCount,
        percent: progressPercent,
      },
      message:
        newStatus === 'finished'
          ? `Congratulations! You finished "${session.book.title}"! 📚🦞 Now write a review to share your thoughts.`
          : newStatus === 'abandoned'
            ? `Marked "${session.book.title}" as DNF (Did Not Finish).`
            : `Progress updated: ${progressPercent.toFixed(1)}% complete.`,
      nextAction:
        newStatus === 'finished'
          ? {
              action: 'Write a review',
              url: '/api/v1/reviews',
              method: 'POST',
            }
          : newStatus === 'reading' && body.currentChunk < session.book.chunkCount
            ? {
                action: 'Read next chunk',
                url: `/api/v1/library/book/${session.book.id}/chunk/${body.currentChunk + 1}`,
                method: 'GET',
              }
            : null,
    };

    return NextResponse.json(response, { headers: rateLimitHeaders });
  } catch (error) {
    console.error('Progress update error:', error);

    if (error instanceof AuthError) {
      return NextResponse.json(
        { error: error.message },
        { status: error.status }
      );
    }

    if (error instanceof RateLimitError) {
      return NextResponse.json(
        { error: error.message },
        { status: 429, headers: error.headers }
      );
    }

    return NextResponse.json(
      { error: 'Failed to update progress' },
      { status: 500 }
    );
  }
}